--- a/src/main/java/org/htmlunit/css/ElementCssStyleDeclaration.java
+++ b/src/main/java/org/htmlunit/css/ElementCssStyleDeclaration.java
@@ -86,7 +86,7 @@ public class ElementCssStyleDeclaration extends AbstractCssStyleDeclaration {
         if (element != null && element.getValue() != null) {
             final String value = element.getValue();
             if (!value.contains("url")) {
-                return StringUtils.toRootLowerCaseWithCache(value);
+                return StringUtils.toRootLowerCase(value);
             }
             return value;
         }