--- a/src/main/java/org/htmlunit/html/DomNamespaceNode.java
+++ b/src/main/java/org/htmlunit/html/DomNamespaceNode.java
@@ -14,12 +14,11 @@
  */
 package org.htmlunit.html;
 
-import java.util.Locale;
-
 import org.htmlunit.SgmlPage;
 import org.htmlunit.WebAssert;
 import org.htmlunit.html.xpath.XPathHelper;
 import org.htmlunit.javascript.host.dom.Document;
+import org.htmlunit.util.StringUtils;
 
 /**
  * Intermediate base class for DOM Nodes that have namespaces. That includes HtmlElement and HtmlAttr.
@@ -61,7 +60,7 @@ public abstract class DomNamespaceNode extends DomNode {
             prefix_ = qualifiedName_.substring(0, colonPosition);
         }
 
-        localNameLC_ = localName_.toLowerCase(Locale.ROOT);
+        localNameLC_ = StringUtils.toRootLowerCase(localName_);
     }
 
     /**