… PERFORMANCE.md
… src/main/java/org/htmlunit/html/HtmlElement.java
… src/main/java/org/htmlunit/html/HtmlImage.java
… src/main/java/org/htmlunit/html/HtmlInput.java
… src/main/java/org/htmlunit/html/HtmlPage.java
… src/main/java/org/htmlunit/html/HtmlRadioButtonInput.java
… src/main/java/org/htmlunit/html/HtmlScript.java
… src/main/java/org/htmlunit/html/HtmlSelect.java
… src/main/java/org/htmlunit/html/HtmlTextArea.java
… src/main/java/org/htmlunit/html/parser/neko/HtmlUnitNekoDOMBuilder.java
… src/main/java/org/htmlunit/html/parser/neko/HtmlUnitNekoHtmlParser.java
… src/main/java/org/htmlunit/javascript/host/Element.java
… src/main/java/org/htmlunit/javascript/host/html/HTMLBodyElement.java
… src/main/java/org/htmlunit/javascript/host/html/HTMLDocument.java
… src/main/java/org/htmlunit/javascript/host/html/HTMLElement.java
… src/main/java/org/htmlunit/javascript/host/html/HTMLSpanElement.java
… src/main/java/org/htmlunit/javascript/host/xml/XMLHttpRequest.java
… src/main/java/org/htmlunit/javascript/host/xml/XMLSerializer.java
… src/main/java/org/htmlunit/svg/SvgElementFactory.java
… src/main/java/org/htmlunit/util/MimeType.java
… src/main/java/org/htmlunit/util/OrderedFastHashMap.java
… src/main/java/org/htmlunit/util/StringUtils.java
… src/main/java/org/htmlunit/util/XmlUtils.java
… src/test/java/org/htmlunit/util/OrderedFastHashMapTest.java
… src/test/java/org/htmlunit/util/RandomUtils.java