HtmlUnit tuning regarding speed and memory. Current success, more on the memory side.
Please run the full test program, there might be something hidden. The small one looked ok.

## E-commerce retailer website

### NEW (JDK 17)
```
Benchmark                                    Mode  Cnt        Score   Error   Units
HtmlUnitBenchmark.simple                     avgt    2  9639101.018           ns/op
HtmlUnitBenchmark.simple:gc.alloc.rate.norm  avgt    2  7612275.912            B/op
```

### OLD (JDK 17)
```
Benchmark                                    Mode  Cnt         Score   Error   Units
HtmlUnitBenchmark.simple                     avgt    2  10527569.765           ns/op
HtmlUnitBenchmark.simple:gc.alloc.rate.norm  avgt    2   9213538.844            B/op
``` 

## XC Homepage

### NEW (JDK 17)
```
Benchmark                                    Mode  Cnt       Score   Error   Units
HtmlUnitBenchmark.simple                     avgt    2  397427.602           ns/op
HtmlUnitBenchmark.simple:gc.alloc.rate.norm  avgt    2  383380.466            B/op
```
### OLD (JDK 17)
```
Benchmark                                    Mode  Cnt       Score   Error   Units
HtmlUnitBenchmark.simple                     avgt    2  411880.438           ns/op
HtmlUnitBenchmark.simple:gc.alloc.rate.norm  avgt    2  454900.410            B/op
```