@rbri 2024-01-28
https://jenkins.wetator.org/job/htmlunit/view/change-requests/job/PR-715/

the failing tests are my ones - i'm still in the middle of a huge refactoring