@rbri 2024-01-28
https://jenkins.wetator.org/job/HtmlUnit%20-%20Library%20Tests/view/change-requests/job/PR-715/

lets see if they also pass