@rbri 2024-01-28
@shapiroronny is this something you are also interested in? Will be great to see if all the hard work from @rschwietzke has some impact for your use case