@shapiroronny 2024-01-28
> @shapiroronny is this something you are also interested in? Will be great to see if all the hard work from @rschwietzke has some impact for your use case

Hopefully yes. We're disabling javascript for our case and don't do any navigations and very little DOM mutation so not sure which flows are optimized. If initial parsing is faster, that's huge for us.

In any case, we haven't bumped in a while so we're happy upgrading in steps in case this was planned to go in the next major version with Computed Css fixes.