@rbri 2024-01-28
> Hopefully yes. We're disabling javascript for our case and don't do any navigations and very little DOM mutation so not sure which flows are optimized. If initial parsing is faster, that's huge for us.

Great; that was my understanding of your use case and of @rschwietzke 's goal.
Will merge that soon and we will see....